Move invoke-dynamic to top level class
